Fix null/zero id bug
This fixes a bug where falsey `id` values would be incorrectly ignored
and treated as notifications.

The JSON-RPC 2.0 spec allows any `number`, `string`, or `null` `id` as
valid, including falsey values like zero or an empty string. This `id`
must be included in the response.

Tests are added to cover this and other test cases around the `id` value.

Fixes #19.